+++ NEWS.md
@@ -2,6 +2,7 @@ v0.1-2 (2015-05-26)
 ===
 
 - proper handling of corner cases (grand total controls and dummy controls with only one category)
+- proper handling of corner cases (grand total controls)
 - explicit error message if reference sample is not sorted
 
 v0.1-1 (2015-05-26)